Improve `git log` parsing
Before this patch, double quotes (") in commit message subjects break
the pseudo-JSON the "--pretty" option's format string tries to wrestle
from `git log`.

With this patch, `git log` is made to emit a newline-delimited
representation of each commit, which is then parsed and (weakly)
validated, and ends up in a dict without intermittent JSONification.